Fix (TypeMap): new SchemaCache sub class
