+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -614,7 +614,6 @@ module ActiveRecord
           self.class.register_class_with_precision m, "timestamp", OID::Timestamp, timezone: @default_timezone
           self.class.register_class_with_precision m, "timestamptz", OID::TimestampWithTimeZone
 
-          # puts "load_additional_types from initialize_type_map"
           load_additional_types
         end
 
@@ -702,7 +701,6 @@ module ActiveRecord
 
         def get_oid_type(oid, fmod, column_name, sql_type = "")
           if !type_map.key?(oid)
-            # puts "load_additional_types from get_oid_type"
             load_additional_types([oid])
           end
 
@@ -721,12 +719,10 @@ module ActiveRecord
           # if schema_cache.additional_type_records.present?
           # TODO: extract this to another method and cover by test
           if should_load_types_from_cache?(oids)
-            # puts "Getting additional types from SC\r\n"
             records = schema_cache.additional_type_records
             initializer.run(records)
           else
-            # puts "Getting additional types from DB\r\n"
-            load_types_queries(initializer, oids) do |query|  
+            load_types_queries(initializer, oids) do |query|
               execute_and_clear(query, "SCHEMA", []) do |records|
                 schema_cache.additional_type_records |= records.to_a
                 initializer.run(records)
@@ -736,8 +732,6 @@ module ActiveRecord
         end
 
         def load_types_queries(initializer, oids)
-          # puts "Getting additional_type_records from DB"
-
           query = <<~SQL
             SELECT t.oid, t.typname, t.typelem, t.typdelim, t.typinput, r.rngsubtype, t.typtype, t.typbasetype
             FROM pg_type as t
@@ -1023,13 +1017,10 @@ module ActiveRecord
           }
 
           if schema_cache.known_coder_type_records.present?
-            # puts "Getting known_coder_type_records from Schema cache\r\n"
-
             coders = schema_cache
               .known_coder_type_records
               .filter_map { |row| construct_coder(row, coders_by_name[row["typname"]]) }
           else
-            # puts "Getting known_coder_type_records from DB\r\n"
             known_coder_types = coders_by_name.keys.map { |n| quote(n) }
 
             query = <<~SQL % known_coder_types.join(", ")
