+++ activerecord/lib/active_record/connection_adapters/schema_cache.rb
@@ -35,18 +35,14 @@ module ActiveRecord
       attr_reader :version
       attr_accessor :connection
 
-      cattr_accessor :additional_type_records, default: []
-      cattr_accessor :known_coder_type_records, default: []
-
       def initialize(conn)
         @connection = conn
+
         @columns      = {}
         @columns_hash = {}
         @primary_keys = {}
         @data_sources = {}
         @indexes      = {}
-        self.additional_type_records = []
-        self.known_coder_type_records = []
       end
 
       def initialize_dup(other)
@@ -67,8 +63,6 @@ module ActiveRecord
         coder["indexes"]          = @indexes
         coder["version"]          = @version
         coder["database_version"] = database_version
-        coder["additional_type_records"] = self.additional_type_records
-        coder["known_coder_type_records"] = self.known_coder_type_records
       end
 
       def init_with(coder)
@@ -78,8 +72,6 @@ module ActiveRecord
         @indexes          = coder["indexes"] || {}
         @version          = coder["version"]
         @database_version = coder["database_version"]
-        self.additional_type_records = coder["additional_type_records"]
-        self.known_coder_type_records = coder["known_coder_type_records"]
 
         derive_columns_hash_and_deduplicate_values
       end
@@ -192,11 +184,11 @@ module ActiveRecord
       def marshal_dump
         reset_version!
 
-        [@version, @columns, {}, @primary_keys, @data_sources, @indexes, database_version, self.known_coder_type_records, self.additional_type_records]
+        [@version, @columns, {}, @primary_keys, @data_sources, @indexes, database_version]
       end
 
       def marshal_load(array)
-        @version, @columns, _columns_hash, @primary_keys, @data_sources, @indexes, @database_version, self.known_coder_type_records, self.additional_type_records = array
+        @version, @columns, _columns_hash, @primary_keys, @data_sources, @indexes, @database_version = array
         @indexes ||= {}
 
         derive_columns_hash_and_deduplicate_values
