+++ activerecord/test/cases/connection_adapters/schema_cache_test.rb
@@ -11,21 +11,6 @@ module ActiveRecord
         @database_version = @connection.get_database_version
       end
 
-      def test_type_map_existence_in_schema_cache
-        # TODO: Check why this test doesn't work
-        return unless current_adapter?(:PostgreSQLAdapter)
-
-        db_config = ActiveRecord::Base.configurations.configs_for(
-          env_name: 'arunit',
-          name: 'primary'
-        )
-
-        connection = ActiveRecord::Base.postgresql_connection(db_config.configuration_hash)
-
-        assert_not connection.schema_cache.additional_type_records.empty?
-        assert_not connection.schema_cache.known_coder_type_records.empty?
-      end
-
       def test_yaml_dump_and_load
         # Create an empty cache.
         cache = SchemaCache.new @connection
@@ -49,13 +34,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            puts cache.additional_type_records.inspect
-            puts cache.inspect
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
       ensure
         tempfile.unlink
@@ -94,11 +72,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
 
         # Load the cache the usual way.
@@ -114,11 +87,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
       ensure
         tempfile.unlink
@@ -233,11 +201,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
       end
 
@@ -260,11 +223,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
       ensure
         tempfile.unlink
@@ -326,11 +284,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
 
         # Load a new cache.
@@ -344,11 +297,6 @@ module ActiveRecord
           assert_equal "id", cache.primary_keys("posts")
           assert_equal 1, cache.indexes("posts").size
           assert_equal @database_version.to_s, cache.database_version.to_s
-
-          if current_adapter?(:PostgreSQLAdapter)
-            assert_not cache.additional_type_records.empty?
-            assert_not cache.known_coder_type_records.empty?
-          end
         end
       ensure
         tempfile.unlink
